토큰 유효기간 추가

+++ src/js/user/logout.js
@@ -1,16 +1,17 @@
 import axios from "axios";
 
 export const Logout = async () => {
+    const token = getTokenWithExpiry();
+    if(token == null) return false;
     try {
         const url = import.meta.env.VITE_API_URL + '/user/logout';
         const res = await axios.get(url, {
             headers: {
-                'Authorization': `Bearer ${localStorage.getItem("token")}`
+                'Authorization': `Bearer ${token}`
             }
         });
 
         if (res.data.result == "success") {
-            alert("로그아웃");
             localStorage.clear();
             return false;
         } else {
@@ -21,4 +22,19 @@ export const Logout = async () => {
     }
 
     return true;
+}
+
+export const getTokenWithExpiry = () => {
+    const strToken = localStorage.getItem('token');
+    if(!strToken) {
+        return null;
+    }
+
+    const token = JSON.parse(strToken);
+    const now = new Date();
+    if(now.getTime() > token.expiry) {
+        localStorage.clear();
+        return null;
+    }
+    return token.value;
 }
